make NetworkInfo usable across :app and :lib
