+++ app/src/main/java/com/alphawallet/app/entity/NetworkInfo.java
@@ -1,37 +1,8 @@
 package com.alphawallet.app.entity;
 
-public class NetworkInfo {
-    public final String name;
-    public final String symbol;
-    public final String rpcServerUrl;
-    public final String etherscanUrl;
-    public final int chainId;
-    public final boolean isMainNetwork;
-    public final String backupNodeUrl;
-    public final String etherscanTxUrl;
-    public final String tickerId;
-    public final String blockscoutAPI;
-
-    public NetworkInfo(
-            String name,
-            String symbol,
-            String rpcServerUrl,
-            String etherscanUrl,
-            int chainId,
-            boolean isMainNetwork,
-            String tickerId,
-            String blockscoutAPI) {
-        this.name = name;
-        this.symbol = symbol;
-        this.rpcServerUrl = rpcServerUrl;
-        this.etherscanUrl = etherscanUrl;
-        this.chainId = chainId;
-        this.isMainNetwork = isMainNetwork;
-        this.backupNodeUrl = null;
-        this.etherscanTxUrl = null;
-        this.tickerId = tickerId;
-        this.blockscoutAPI = blockscoutAPI;
-    }
+public class NetworkInfo extends com.alphawallet.ethereum.NetworkInfo {
+    public String backupNodeUrl = null;
+    public String etherscanTxUrl = null;
 
     public NetworkInfo(
             String name,
@@ -44,16 +15,9 @@ public class NetworkInfo {
             String etherscanTxUrl,
             String tickerId,
             String blockscoutAPI) {
-        this.name = name;
-        this.symbol = symbol;
-        this.rpcServerUrl = rpcServerUrl;
-        this.etherscanUrl = etherscanUrl;
-        this.chainId = chainId;
-        this.isMainNetwork = isMainNetwork;
+        super(name, symbol, rpcServerUrl, etherscanUrl, chainId, isMainNetwork, tickerId, blockscoutAPI);
         this.backupNodeUrl = backupNodeUrl;
         this.etherscanTxUrl = etherscanTxUrl;
-        this.tickerId = tickerId;
-        this.blockscoutAPI = blockscoutAPI;
     }
 
     public String getShortName()
